requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
fonttools==4.54.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.3
